+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This library _tries_ to adhere to [Semantic Versioning](http://semver.org/).
 - added new selectors FVSelectionBy (Box,Sphere,Cylinder) suitable for MimmoFvMesh objects.
 - added reading and writing STL solid names in multiSolid STL.
 - added MeshChecker class  in "utils" module
+- added SurfaceTriangulator class to handle mixed type 3D surface tessellation as homogeneous triangulations.
 
 ### Changed
 - PropagateField classes: solution of Laplacian system solved by using bitpit LA module.